complete the functopn of save and rebuilt
S for save, B for build and C for clear.
The objects can now be only be rebuilt after the scene has been switched due to the naming issue(can be fixed by using prefab) . Another improvement can be made by giving user a reminder when data is saved and cleared.